pyproject.toml
[tool.poetry]
name = "chain-factory"
version = "0.1.0"
description = "A CLI utitlity to mass manufacture LLM chains."
authors = ["Pankaj Garkoti <garkotipankaj@gmail.com>"]
readme = "README.md"
packages = [{include = "ChainFactory"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
